Removed code that required person code to be positive in the data access layer. The authorization code makes sure that there is a valid person code when entering the data layer but also the person code is now an unsigned integer which means the filter on person code will always work in the sql.

+++ DataAccessLayer/Repository/IPhenomenaDataAccess.cs
@@ -6,7 +6,7 @@ namespace DataAccessLayer
 {
     public interface IPhenomenaDataAccess
     {
-        IEnumerable<Phenomena> GetPhenomena(long personID);
+        IEnumerable<Phenomena> GetPhenomena(uint personID);
 
         //TODO Implement
         //long InsertNote(long personID, DateTime date, string note, bool behaviorChange);
