+++ DataAccessLayer/Repository/IPrescriptionDataAccess.cs
@@ -7,7 +7,7 @@ namespace DataAccessLayer
     public interface IPrescriptionDataAccess
     {      
 
-        IEnumerable<DataAccessLayer.Models.Prescription> GetPrescriptions(long personID);
+        IEnumerable<DataAccessLayer.Models.Prescription> GetPrescriptions(uint personID);
         
     }
 }
