+++ DataAccessLayer/Repository/ITableNotesLinksDataAccess.cs
@@ -6,7 +6,7 @@ namespace DataAccessLayer
 {
     public interface ITableNotesLinksDataAccess
     {
-        long Insert(long personID, long[] noteIDs, string table_name, long entity_id);    
+        long Insert(uint personID, long[] noteIDs, string table_name, long entity_id);    
         
     }
 }
